requirements.txt
pytest-django
django-debug-toolbar
pytest-cov
factory-boy
pandas
robin-stocks
numpy
django
django-environ
djangorestframework
django-model-utils
argon2-cffi
django-allauth
django-crispy-forms
django-extensions
coreapi
python-twitter
psycopg2-binary
PyMySQL==0.9.3
google-cloud-ndb
